+++ style_generator/lib/style_generator.dart
@@ -3,6 +3,7 @@ import "package:build/build.dart";
 import "src/builder/copy_with/copy_with_builder.dart";
 import "src/builder/equality/equality_builder.dart";
 import "src/builder/style/style_builder.dart";
+import "src/builder/to_string/to_string_builder.dart";
 
 export "src/builder/copy_with/copy_with_builder.dart";
 export "src/builder/style/style_builder.dart";
@@ -12,3 +13,5 @@ Builder styleBuilder(BuilderOptions options) => StyleBuilder(options: options);
 Builder copyWithBuilder(BuilderOptions options) => CopyWithBuilder(options: options);
 
 Builder equalityBuilder(BuilderOptions options) => EqualityBuilder(options: options);
+
+Builder toStringBuilder(BuilderOptions options) => ToStringBuilder(options: options);