+++ style_generator_annotation/lib/to_string_generator_annotation.dart
@@ -1,0 +1,2 @@
+export "src/annotations/to_string.dart";
+export "src/annotations/to_string_key.dart";